refactor(ui): restructure folders
