+++ apps/ui/src/components/Trace/Editor/query-viewer/QueryType.tsx
@@ -1,6 +1,6 @@
 import { Kind, TypeNode } from "graphql";
 
-export function Type({ type }: { type: TypeNode }) {
+export function QueryType({ type }: { type: TypeNode }) {
   if (type.kind === Kind.NAMED_TYPE) {
     return <span>{type.name.value}</span>;
   }
@@ -8,7 +8,7 @@ export function Type({ type }: { type: TypeNode }) {
   if (type.kind === Kind.LIST_TYPE) {
     return (
       <span>
-        [<Type type={type.type} />]
+        [<QueryType type={type.type} />]
       </span>
     );
   }
@@ -16,7 +16,7 @@ export function Type({ type }: { type: TypeNode }) {
   if (type.kind === Kind.NON_NULL_TYPE) {
     return (
       <span>
-        <Type type={type.type} />!
+        <QueryType type={type.type} />!
       </span>
     );
   }
