+++ apps/ui/src/components/index.ts
@@ -1,4 +1,5 @@
 export * from "./SideBar";
 export * from "./SchemaOverview/ListSchema";
 export * from "./SchemaOverview/NoSchema";
+export * from "./Trace";
 export * from "./Schema";
