+++ apps/ui/src/pages/Schema.tsx
@@ -3,9 +3,7 @@ import { useState } from "react";
 import { useParams } from "react-router-dom";
 
 import { listSchemas } from "../../src/api/list-schemas";
-import { SideBar } from "../components";
-import { SchemaViewer } from "../components/Schema/SchemaViewer";
-import { Trace } from "../components/Trace";
+import { SchemaViewer, SideBar, Trace } from "../components";
 
 export const Schema = () => {
   const [displaySchema, setDisplaySchema] = useState(true);
